pyproject.toml
[project]
name = "ontoenv_pirate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.60.2",
    "python-dotenv>=1.0.1",
    "rdflib>=7.1.3",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
